Refactors and bugfixes from `main` branch

+++ app/src/main/java/uk/openvk/android/legacy/api/wrappers/DownloadManager.java
@@ -5,6 +5,8 @@ import android.content.Context;
 import android.content.pm.PackageInfo;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Looper;
 import android.os.Message;
 import android.util.Log;
 
@@ -33,8 +35,12 @@ import okhttp3.Response;
 import uk.openvk.android.legacy.BuildConfig;
 import uk.openvk.android.legacy.OvkApplication;
 import uk.openvk.android.legacy.api.attachments.PhotoAttachment;
+import uk.openvk.android.legacy.api.interfaces.OvkAPIListeners;
 import uk.openvk.android.legacy.ui.core.activities.AppActivity;
 import uk.openvk.android.legacy.ui.core.activities.AuthActivity;
+import uk.openvk.android.legacy.ui.core.activities.base.NetworkActivity;
+import uk.openvk.android.legacy.ui.core.activities.base.NetworkAuthActivity;
+import uk.openvk.android.legacy.ui.core.activities.base.NetworkFragmentActivity;
 import uk.openvk.android.legacy.ui.core.activities.intents.GroupIntentActivity;
 import uk.openvk.android.legacy.ui.core.activities.PhotoViewerActivity;
 import uk.openvk.android.legacy.ui.core.activities.QuickSearchActivity;
@@ -73,8 +79,15 @@ public class DownloadManager {
     private HttpClient httpClientLegacy = null;
     private boolean forceCaching;
     private String instance;
+    OvkAPIListeners apiListeners;
+    Handler handler;
 
-    public DownloadManager(Context ctx, boolean use_https, boolean legacy_mode) {
+    public DownloadManager(Context ctx, boolean use_https, boolean legacy_mode, Handler handler) {
+        this.handler = handler;
+        apiListeners = new OvkAPIListeners();
+        if(handler == null) {
+            searchHandler();
+        }
         this.ctx = ctx;
         this.use_https = use_https;
         this.legacy_mode = legacy_mode;
@@ -117,12 +130,20 @@ public class DownloadManager {
                     sslContext.init(null, trustAllCerts, new java.security.SecureRandom());
                     javax.net.ssl.SSLSocketFactory ssf = (javax.net.ssl.SSLSocketFactory)
                             sslContext.getSocketFactory();
-                    httpClient = new OkHttpClient.Builder().sslSocketFactory(sslContext.getSocketFactory())
-                            .connectTimeout(30, TimeUnit.SECONDS).writeTimeout(30, TimeUnit.SECONDS)
-                            .readTimeout(30, TimeUnit.SECONDS).retryOnConnectionFailure(false).build();
+                    httpClient = new OkHttpClient.Builder()
+                            .sslSocketFactory(sslContext.getSocketFactory())
+                            .connectTimeout(30, TimeUnit.SECONDS)
+                            .writeTimeout(30, TimeUnit.SECONDS)
+                            .readTimeout(30, TimeUnit.SECONDS)
+                            .retryOnConnectionFailure(false)
+                            .build();
                 } catch (Exception e) {
-                    httpClient = new OkHttpClient.Builder().connectTimeout(30, TimeUnit.SECONDS).writeTimeout
-                            (30, TimeUnit.SECONDS).readTimeout(30, TimeUnit.SECONDS).retryOnConnectionFailure(false).build();
+                    httpClient = new OkHttpClient.Builder()
+                            .connectTimeout(30, TimeUnit.SECONDS)
+                            .writeTimeout(30, TimeUnit.SECONDS)
+                            .readTimeout(30, TimeUnit.SECONDS)
+                            .retryOnConnectionFailure(false)
+                            .build();
                 }
                 legacy_mode = false;
             }
@@ -231,9 +252,10 @@ public class DownloadManager {
                                 + " ms | Filesize: " + downloadedFile.length() + " bytes");
                     } else if (photoAttachment.url.length() == 0) {
                         filename = photoAttachment.filename;
-                        if(logging_enabled) Log.e(OvkApplication.DL_TAG, "Invalid address. Skipping...");
+                        if(logging_enabled) Log.e(OvkApplication.DL_TAG,
+                                "Invalid or empty URL. Skipping...");
                         try {
-                            if(downloadedFile.exists() || !downloadedFile.isDirectory()) {
+                            if(downloadedFile.exists() && !downloadedFile.isDirectory()) {
                                 FileOutputStream fos = new FileOutputStream(downloadedFile);
                                 byte[] bytes = new byte[1];
                                 bytes[0] = 0;
@@ -256,7 +278,8 @@ public class DownloadManager {
                             // short_address, i + 1, photoAttachments.size()));
                             url = photoAttachments.get(i).url;
                             if(!url.startsWith("http://") && !url.startsWith("https://")) {
-                                Log.e(OvkApplication.DL_TAG, "Invalid URL. Download canceled.");
+                                Log.e(OvkApplication.DL_TAG,
+                                        String.format("Invalid URL: %s. Download canceled.", url));
                                 return;
                             }
 
@@ -341,6 +364,9 @@ public class DownloadManager {
                             case "newsfeed_avatars":
                                 sendMessage(HandlerMessages.NEWSFEED_AVATARS, where);
                                 break;
+                            case "photo_albums":
+                                sendMessage(HandlerMessages.PHOTO_ALBUM_THUMBNAILS, where);
+                                break;
                             case "group_avatars":
                                 sendMessage(HandlerMessages.GROUP_AVATARS, where);
                                 break;
@@ -361,6 +387,9 @@ public class DownloadManager {
                                 break;
                             case "comment_photos":
                                 sendMessage(HandlerMessages.COMMENT_PHOTOS, where);
+                                break;
+                            case "album_photos":
+                                sendMessage(HandlerMessages.ALBUM_PHOTOS, where);
                                 break;
                             case "conversations_avatars":
                                 sendMessage(HandlerMessages.CONVERSATIONS_AVATARS, where);
@@ -385,7 +414,7 @@ public class DownloadManager {
             return;
         }
         if(!url.startsWith("http://") && !url.startsWith("https://")) {
-            Log.e(OvkApplication.DL_TAG, "Invalid URL. Download canceled.");
+            Log.e(OvkApplication.DL_TAG, String.format("Invalid URL: %s. Download canceled.", url));
             return;
         }
         Runnable httpRunnable = new Runnable() {
@@ -446,7 +475,8 @@ public class DownloadManager {
                         short_address = url;
                     }
 
-                    if(logging_enabled) Log.v("DownloadManager", String.format("Downloading %s...", short_address));
+                    if(logging_enabled) Log.v("DownloadManager",
+                            String.format("Downloading %s...", short_address));
                     if (legacy_mode) {
                         request_legacy = httpClientLegacy.get(url);
                     } else {
@@ -540,6 +570,9 @@ public class DownloadManager {
                     case "comment_photos":
                         sendMessage(HandlerMessages.COMMENT_PHOTOS, where);
                         break;
+                    case "album_photos":
+                        sendMessage(HandlerMessages.ALBUM_PHOTOS, where);
+                        break;
                     case "video_thumbnails":
                         sendMessage(HandlerMessages.VIDEO_THUMBNAILS, where);
                         break;
@@ -554,50 +587,50 @@ public class DownloadManager {
         thread.start();
     }
 
-    private void sendMessage(int message, String response) {
+    private void sendMessage(final int message, String response) {
         Message msg = new Message();
         msg.what = message;
-        Bundle bundle = new Bundle();
+        final Bundle bundle = new Bundle();
         bundle.putString("response", response);
         msg.setData(bundle);
-        if(ctx.getClass().getSimpleName().equals("AuthActivity")) {
-            ((AuthActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("AppActivity")) {
-            ((AppActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("QuickSearchActivity")) {
-            ((QuickSearchActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("ProfileIntentActivity")) {
-            ((ProfileIntentActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("FriendsIntentActivity")) {
-            ((FriendsIntentActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("GroupIntentActivity")) {
-            ((GroupIntentActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("WallPostActivity")) {
-            ((WallPostActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("PhotoViewerActivity")) {
-            ((PhotoViewerActivity) ctx).handler.sendMessage(msg);
-        }
+        handler.post(new Runnable() {
+            @Override
+            public void run() {
+                if(message < 0) {
+                    apiListeners.failListener.onAPIFailed(ctx, message, bundle);
+                } else {
+                    apiListeners.successListener.onAPISuccess(ctx, message, bundle);
+                }
+            }
+        });
     }
 
-    private void sendMessage(int message, String response, int id) {
+    private void sendMessage(final int message, String response, int id) {
         Message msg = new Message();
         msg.what = message;
-        Bundle bundle = new Bundle();
+        final Bundle bundle = new Bundle();
         bundle.putString("response", response);
         bundle.putInt("id", id);
         msg.setData(bundle);
-        if(ctx.getClass().getSimpleName().equals("AuthActivity")) {
-            ((AuthActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("AppActivity")) {
-            ((AppActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("ProfileIntentActivity")) {
-            ((ProfileIntentActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("FriendsIntentActivity")) {
-            ((FriendsIntentActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("GroupIntentActivity")) {
-            ((GroupIntentActivity) ctx).handler.sendMessage(msg);
-        } else if(ctx.getClass().getSimpleName().equals("CommentsIntentActivity")) {
-            ((WallPostActivity) ctx).handler.sendMessage(msg);
+        handler.post(new Runnable() {
+            @Override
+            public void run() {
+                if(message < 0) {
+                    apiListeners.failListener.onAPIFailed(ctx, message, bundle);
+                } else {
+                    apiListeners.successListener.onAPISuccess(ctx, message, bundle);
+                }
+            }
+        });
+    }
+
+    private void searchHandler() {
+        if(ctx instanceof NetworkFragmentActivity) {
+            this.handler = ((NetworkFragmentActivity) ctx).handler;
+        } else if(ctx instanceof NetworkAuthActivity) {
+            this.handler = ((NetworkAuthActivity) ctx).handler;
+        } else if(ctx instanceof NetworkActivity) {
+            this.handler = ((NetworkActivity) ctx).handler;
         }
     }
 
@@ -650,5 +683,9 @@ public class DownloadManager {
         };
         new Thread(runnable).run();
         return size[0];
+    }
+
+    public void setAPIListeners(OvkAPIListeners apiListeners) {
+        this.apiListeners = apiListeners;
     }
 }
